Fix linter warnings in util.c

+++ src/util.c
@@ -10,12 +10,13 @@
 #include <strings.h>
 #include <stdlib.h>
 #include <ctype.h>
-#include <assert.h>
+
 #include "redisraft.h"
 
 int RedisModuleStringToInt(RedisModuleString *str, int *value)
 {
     long long tmpll;
+
     if (RedisModule_StringToLongLong(str, &tmpll) != REDISMODULE_OK) {
         return REDISMODULE_ERR;
     }
@@ -24,7 +25,7 @@ int RedisModuleStringToInt(RedisModuleString *str, int *value)
         return REDISMODULE_ERR;
     }
 
-    *value = tmpll;
+    *value = (int) tmpll;
     return REDISMODULE_OK;
 }
